Adaptive result streaming implementation
This handles the situation of having too many results sent over
a network that is too slow, but should work equally well for a fast
network.

This implements per-client logic that "rolls up" multiple updates
into one. Combined with the backpressure from the websocket connection,
this means we only send as fast as the connection supports. In addition,
we also don't block the `acquisition_loop`, which can run independently
from sending the results to clients.

Also disable compression, because we do this ourselves.